Added new helper for encrypt/decrypt and embedded data uri. Updated MenuEntry for data uri and ssl urls. Version_check is removed, in favor of a single bootstrap, now localed in main /lib.

+++ lib/Pf4wp/Options/WordpressOptions.php
@@ -57,7 +57,8 @@ class WordpressOptions extends Options
      */
     public function delete()
     {
-        parent::delete();
+        if (!parent::delete())
+            return false;
         
         $result = delete_option($this->name);
